Split out Result and ResultType to break import cycle
Previously, there was an import cycle between tui.menu_item and
tui.types.
