+++ archinstall/lib/disk/encryption_menu.py
@@ -13,7 +13,8 @@ from archinstall.lib.models.device_model import (
 )
 from archinstall.tui.curses_menu import SelectMenu
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
-from archinstall.tui.types import Alignment, FrameProperties, ResultType
+from archinstall.tui.result import ResultType
+from archinstall.tui.types import Alignment, FrameProperties
 
 from ..menu.abstract_menu import AbstractSubMenu
 from ..models.device_model import Fido2Device
