+++ archinstall/lib/disk/partitioning_menu.py
@@ -19,7 +19,8 @@ from archinstall.lib.models.device_model import (
 )
 from archinstall.tui.curses_menu import EditMenu, SelectMenu
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
-from archinstall.tui.types import Alignment, FrameProperties, Orientation, ResultType
+from archinstall.tui.result import ResultType
+from archinstall.tui.types import Alignment, FrameProperties, Orientation
 
 from ..menu.list_manager import ListManager
 from ..output import FormattedOutput
