+++ archinstall/lib/disk/subvolume_menu.py
@@ -3,7 +3,8 @@ from typing import TYPE_CHECKING, assert_never, override
 
 from archinstall.lib.models.device_model import SubvolumeModification
 from archinstall.tui.curses_menu import EditMenu
-from archinstall.tui.types import Alignment, ResultType
+from archinstall.tui.result import ResultType
+from archinstall.tui.types import Alignment
 
 from ..menu.list_manager import ListManager
 from ..utils.util import prompt_dir
