+++ archinstall/lib/interactions/manage_users_conf.py
@@ -5,7 +5,8 @@ from typing import TYPE_CHECKING, override
 
 from archinstall.tui.curses_menu import EditMenu, SelectMenu
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
-from archinstall.tui.types import Alignment, Orientation, ResultType
+from archinstall.tui.result import ResultType
+from archinstall.tui.types import Alignment, Orientation
 
 from ..menu.list_manager import ListManager
 from ..models.users import User
