+++ archinstall/lib/locale/locale_menu.py
@@ -2,7 +2,8 @@ from typing import TYPE_CHECKING, override
 
 from archinstall.tui.curses_menu import SelectMenu
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
-from archinstall.tui.types import Alignment, FrameProperties, ResultType
+from archinstall.tui.result import ResultType
+from archinstall.tui.types import Alignment, FrameProperties
 
 from ..menu.abstract_menu import AbstractSubMenu
 from ..models.locale import LocaleConfiguration
