+++ archinstall/lib/menu/list_manager.py
@@ -3,7 +3,8 @@ from typing import TYPE_CHECKING, Any
 
 from archinstall.tui.curses_menu import SelectMenu
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
-from archinstall.tui.types import Alignment, ResultType
+from archinstall.tui.result import ResultType
+from archinstall.tui.types import Alignment
 
 from ..output import FormattedOutput
