+++ archinstall/tui/__init__.py
@@ -1,6 +1,7 @@
 from .curses_menu import EditMenu, SelectMenu, Tui
 from .menu_item import MenuItem, MenuItemGroup
-from .types import Alignment, Chars, FrameProperties, FrameStyle, Orientation, PreviewStyle, Result, ResultType
+from .result import Result, ResultType
+from .types import Alignment, Chars, FrameProperties, FrameStyle, Orientation, PreviewStyle
 
 __all__ = [
     'Alignment',
